pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "RAG-based Chatbot"
authors = [
    {name = "HaninIsmail01",email = "haninmonir01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "logging (>=0.4.9.6,<0.5.0.0)",
    "qdrant-client (>=1.17.1,<2.0.0)",
    "fastembed (>=0.8.0,<0.9.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "llama-index (>=0.14.20,<0.15.0)",
    "llama-index-vector-stores-qdrant (>=0.10.0,<0.11.0)",
    "llama-index-embeddings-fastembed (>=0.6.0,<0.7.0)",
    "llama-cloud (>=2.1)",
    "flagembedding (>=1.2.0)",
    "sentence-transformers (>=5.4.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
